+++ tgui/packages/tgui/interfaces/Who/constants.ts
@@ -1,0 +1,2 @@
+export const NEW_ACCOUNT_AGE = 7;
+export const NEW_ACCOUNT_NOTICE = 'Аккаунт создан недавно';
