+++ tgui/packages/tgui/styles/interfaces/Who.scss
@@ -1,38 +1,44 @@
+@use 'sass:color';
 @use '../base.scss' as base;
+
+$border-color: color.change(base.$color-bg, $lightness: 20%);
 
 .Who_Table {
   margin: -1px;
-  border: 1px solid hsla(0, 0%, 100%, 0.2);
+  border: 1px solid $border-color;
 
   .Table__row--header {
     position: sticky;
     top: -0.75em;
-    background-color: base.$color-bg;
+    box-shadow: 0px 1px 0 $border-color;
     z-index: 1;
 
-    &:after {
-      content: '';
-      position: absolute;
-      top: 0;
-      right: 0;
-      bottom: 0;
-      left: 0;
-      background-color: base.$color-bg-section;
-      z-index: -2;
-    }
-
     td {
+      position: relative;
       padding: 0.33em 0.5em;
-      border-right: 1px solid hsla(0, 0%, 100%, 0.3);
+      background-color: base.$color-bg;
+      border-right: 1px solid $border-color;
+      z-index: 0;
 
       &:last-child {
         border-right: none;
+      }
+
+      &:after {
+        content: '';
+        position: absolute;
+        top: 0;
+        right: 0;
+        bottom: 0;
+        left: 0;
+        background-color: base.$color-bg-section;
+        z-index: -2;
       }
     }
   }
 
   tr {
-    border-bottom: 1px solid hsla(0, 0%, 100%, 0.2);
+    border-bottom: 1px solid $border-color;
 
     &:last-child {
       border-bottom: none;
@@ -40,7 +46,7 @@
 
     td {
       padding: 0.33em 0.5em !important;
-      border-right: 1px solid hsla(0, 0%, 100%, 0.2);
+      border-right: 1px solid base.$color-bg;
 
       &:last-child {
         border-right: none;
